pyproject.toml
[tool.poetry]
name = "wrap_project"
version = "0.1.0"
description = ""
authors = ["ticochuck <chucklivilla@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
jupyterlab = "^3.0.12"
pandas = "^1.2.3"
django = "^3.2"
django-environ = "^0.4.5"
django-import-export = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
